requirements.txt
langchain
langchain-google-genai
python-dotenv
pillow
streamlit
langchain-chroma